feat: added config-based resource tables
