+++ src/magniFHIR/Config/FhirServersOptions.cs
@@ -16,16 +16,6 @@ public class FhirServerConfig
     public Uri? BaseUrl { get; init; }
     public string? NameSlug => Slugify(Name);
     public FhirServerAuthConfig? Auth { get; set; }
-    public IDictionary<ResourceType, ResourceBrowserConfig> ResourceBrowsers { get; set; }
-
-    public FhirServerConfig()
-    {
-        ResourceBrowsers = new Dictionary<ResourceType, ResourceBrowserConfig>();
-        foreach (var resourceType in Enum.GetValues<ResourceType>())
-        {
-            ResourceBrowsers.Add(resourceType, new ResourceBrowserConfig());
-        }
-    }
 
     public override string ToString()
     {
@@ -89,9 +79,4 @@ public class FhirServerBasicAuthConfig
 {
     public string? Username { get; set; }
     public string? Password { get; set; }
-}
-
-public class ResourceBrowserConfig
-{
-    public int PageSize { get; set; } = 50;
 }
