+++ src/magniFHIR/Pages/PatientBrowser/PatientBrowser.razor
@@ -8,7 +8,7 @@
 
 <MudText Typo="Typo.h3" GutterBottom="true">Patient Browser</MudText>
 
-<MudTable Dense="true" Hover="true" Breakpoint="Breakpoint.Sm" LoadingProgressColor="Color.Info" ServerData="@(new Func<TableState, Task<TableData<Patient>>>(ServerReload))" >
+<MudTable Dense="true" Hover="true" Breakpoint="Breakpoint.Sm" LoadingProgressColor="Color.Info" ServerData="@(new Func<TableState, Task<TableData<Patient>>>(ServerReload))">
     <HeaderContent>
         <MudTh>Name</MudTh>
         <MudTh>Birth Date</MudTh>
@@ -27,22 +27,22 @@
             <MudTd DataLabel="LastUpdated">@context.Meta?.LastUpdated</MudTd>
             <MudTd>
                 <MudButton Size="Size.Small" Variant="Variant.Outlined" Color="Color.Secondary"
-                    EndIcon="@Icons.Material.Filled.ContactPage"
-                    Href="@($"/servers/{FhirServerNameSlug}/patients/{context.Id}")" Target="_blank">
+                           EndIcon="@Icons.Material.Filled.ContactPage"
+                           Href="@($"/servers/{FhirServerNameSlug}/patients/{context.Id}")" Target="_blank">
                     View Record
                 </MudButton>
             </MudTd>
-    </RowTemplate>
-    <NoRecordsContent>
-        <MudText>No matching records found</MudText>
-    </NoRecordsContent>
-    <LoadingContent>
-        <MudText>Loading...</MudText>
-    </LoadingContent>
-    <PagerContent>
-        <MudTablePager HideRowsPerPage="true" HidePageNumber="true" LastIcon="null" FirstIcon="null" BeforeIcon="null"/>
-    </PagerContent>
-</MudTable>
+        </RowTemplate>
+        <NoRecordsContent>
+            <MudText>No matching records found</MudText>
+        </NoRecordsContent>
+        <LoadingContent>
+            <MudText>Loading...</MudText>
+        </LoadingContent>
+        <PagerContent>
+            <MudTablePager HideRowsPerPage="true" HidePageNumber="true" LastIcon="null" FirstIcon="null" BeforeIcon="null" />
+        </PagerContent>
+    </MudTable>
 
 @code {
     [Parameter]
@@ -70,6 +70,6 @@
             totalItems = patients.Count;
         }
 
-        return new TableData<Patient>() {Items = patients, TotalItems = totalItems};
+        return new TableData<Patient>() { Items = patients, TotalItems = totalItems };
     }
 }
