+++ src/magniFHIR/appsettings.Development.json
@@ -11,12 +11,7 @@
   "FhirServers": [
     {
       "Name": "HAPI FHIR JPA Server",
-      "BaseUrl": "http://hapi-fhir-server.127.0.0.1.nip.io/fhir",
-      "ResourceBrowsers": {
-        "Patient": {
-          "PageSize": 3
-        }
-      }
+      "BaseUrl": "http://hapi-fhir-server.127.0.0.1.nip.io/fhir"
     },
     {
       "Name": "Linux4Health FHIR Server (empty)",
@@ -33,6 +28,11 @@
       "BaseUrl": "http://azure-fhir-server.127.0.0.1.nip.io/"
     }
   ],
+  "ResourceBrowsers": {
+    "Patient": {
+      "PageSize": 3
+    }
+  },
   "Tracing": {
     "Enabled": true,
     "Exporter": "jaeger",
