+++ src/magniFHIR/appsettings.json
@@ -9,6 +9,13 @@
   },
   "AllowedHosts": "*",
   "FhirServers": [],
+  "ResourceBrowsers": {
+    "Condition": {
+      "PageSize": 3
+    },
+    "Patient": {
+    }
+  },
   "Tracing": {
     "Enabled": false,
     "ServiceName": "magniFHIR",
